temp

+++ includes/figure.h
@@ -6,7 +6,7 @@
 /*   By: dmitry <dmitry@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/11 20:49:04 by lorphan           #+#    #+#             */
-/*   Updated: 2022/03/21 01:46:42 by dmitry           ###   ########.fr       */
+/*   Updated: 2022/03/22 18:09:08 by dmitry           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@
 
 typedef struct s_pos
 {
-	double	x;
-	double	y;
-	double	z;
+	float	x;
+	float	y;
+	float	z;
 }	t_pos;
 
 typedef struct s_pos	t_vec3;
@@ -36,12 +36,12 @@ typedef struct s_color
 typedef struct s_ambient
 {
 	t_color	color;
-	double	lighting_ratio;
+	float	lighting_ratio;
 }	t_ambient;
 
 typedef struct s_camera
 {
-	double	fov;
+	float	fov;
 	t_pos	pos;
 	t_pos	normal;
 }	t_camera;
@@ -50,14 +50,14 @@ typedef struct s_light
 {
 	t_pos	pos;
 	t_color	color;
-	double	brightness_ratio;
+	float	brightness_ratio;
 }	t_light;
 
 typedef struct s_sphere
 {
 	t_pos	pos;
 	t_color	color;
-	double	diameter;
+	float	diameter;
 }	t_sphere;
 
 typedef struct s_plane
@@ -72,8 +72,8 @@ typedef struct s_cylinder
 	t_pos	pos;
 	t_color	color;
 	t_pos	normal;
-	double	diameter;
-	double	height;
+	float	diameter;
+	float	height;
 }	t_cylinder;
 
 void	print_figure(void *data, int type);
